// Electrohuila - copia/pqr-scheduling-appointments-portal/src/features/admin/models/admin.models.ts
export type TabType = 'citas' | 'empleados' | 'roles' | 'sedes' | 'tipos-cita' | 'horas-disponibles' | 'permisos' | 'festivos' | 'plantillas' | 'settings';